Displays diamonds next to moderators
That way diamonds are displayed everywhere, instead of just in the sidebar. In addition, it adds extra distinction between mods and users. People can imitate being a moderator by having a gravatar with a gold border (they can't add a diamond next to their username), and this will be useful if there's a situation where usernames need to be displayed without avatars.